refactor: move getMessagePage to be a factory
