+++ web/src/gamesShared/components/fbg/GameLayout.test.tsx
@@ -3,7 +3,7 @@ import { mount } from 'enzyme';
 import { GameLayout } from './GameLayout';
 import { IGameArgs } from 'gamesShared/definitions/game';
 import { GameMode } from 'gamesShared/definitions/mode';
-import { LobbyService } from '../../../infra/common/services/LobbyService';
+import { LobbyService } from 'infra/common/services/LobbyService';
 import ReplayIcon from '@material-ui/icons/Replay';
 import { Router } from 'infra/i18n';
