+++ web/src/infra/common/components/alert/LoadingMessage.tsx
@@ -1,7 +1,7 @@
-import MessagePageClass from 'infra/common/components/alert/MessagePageClass';
+import MessagePage from 'infra/common/components/alert/MessagePage';
 import { useTranslation } from 'infra/i18n';
 
 export function LoadingMessage() {
   const { t } = useTranslation('LoadingMessage');
-  return <MessagePageClass type="loading" message={t('loading')} />;
+  return <MessagePage type="loading" message={t('loading')} />;
 }
