+++ web/src/infra/lobby/NewRoomModal.tsx
@@ -6,7 +6,7 @@ import { OccupancySelect } from 'infra/common/components/game/OccupancySelect';
 import NicknameRequired from 'infra/common/components/auth/NicknameRequired';
 import { LobbyService } from 'infra/common/services/LobbyService';
 import { Router, withTranslation, WithTranslation } from 'infra/i18n';
-import getMessagePage from 'infra/common/components/alert/MessagePage';
+import getMessagePage from 'infra/common/factories/MessagePage';
 import { IGameDef } from 'gamesShared/definitions/game';
 import { GamePickerModal } from 'infra/common/components/game/GamePickerModal';
 import { room } from 'infra/navigation';
